requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-tailwind==4.2.0
django-cotton==2.1.3
django-environ==0.12.0
argon2-cffi==25.1.0
python-dotenv==1.0.1
psycopg2==2.9.10
djangorestframework-simplejwt==5.3.1
google-generativeai==0.8.3
django-redis==5.4.0
redis==5.0.4
django-ratelimit==4.1.0
